feat: Add button class and modify two classes
Changes in this commit

A new button class for displaying a button on the screen with a particular
messgae. The showing of this button is triggered by some specific
events, which are determined in game_functions.py.

It adds to game_functions.py the different methods that control the
aliens, the game scoring, and the new button class.

Modifies the game loop to bring everything to working together.